Moved Harvest-specific page into the datastore module. Moved tests
