+++ modules/harvest/modules/harvest_dashboard/tests/src/Unit/Controller/ControllerTest.php
@@ -74,80 +74,6 @@ class ControllerTest extends TestCase {
     }
   }
 
-  public function testNoDatasets() {
-    $time = time();
-
-    $container = $this->getCommonMockChain()
-      ->add(Harvest::class, 'getAllHarvestRunInfo', [$time])
-      ->add(DatasetInfo::class, 'gather', ['notice' => 'Not found'])
-      ->getMock();
-
-    \Drupal::setContainer($container);
-
-    $controller = new Controller();
-    $response = $controller->harvestDatasets('test');
-
-    $json = json_encode($response);
-    $strings = array_merge(Controller::DATASET_HEADERS,);
-
-    $this->assertEmpty($response['#rows']);
-    foreach ($strings as $string) {
-      $this->assertStringContainsString($string, $json);
-    }
-  }
-
-  public function testDataset() {
-    $time = time();
-
-    $info = [
-      'uuid' => 'dataset-1',
-      'title' => 'Title',
-      'revision_id' => '2',
-      'moderation_state' => 'published',
-      'modified_date_metadata' => '2020-01-15',
-      'modified_date_dkan' => '2021-02-11',
-    ];
-
-    $distribution = [
-      'distribution_uuid' => 'dist-1',
-      'fetcher_status' => 'done',
-      'fetcher_percent_done' => 100,
-      'importer_status' => 'done',
-      'importer_percent_done' => 100,
-    ];
-
-    $infoWithDist = array_merge($info, ['distributions' => [$distribution]]);
-
-    $datasetInfo = [
-      'latest_revision' => $infoWithDist,
-    ];
-
-    $container = $this->getCommonMockChain()
-      ->add(Harvest::class, 'getAllHarvestRunInfo', [$time])
-      ->add(DatasetInfo::class, 'gather', $datasetInfo)
-      ->getMock();
-
-    \Drupal::setContainer($container);
-
-    $controller = new Controller();
-    $response = $controller->harvestDatasets('test');
-
-    $json = json_encode($response);
-    $strings = array_merge(
-      Controller::DATASET_HEADERS,
-      Controller::DISTRIBUTION_HEADERS,
-      $info,
-      $distribution
-    );
-
-    foreach ($strings as $string) {
-      $this->assertStringContainsString($string, $json);
-    }
-
-    $title = (string) $controller->harvestDatasetsTitle('test');
-    $this->assertEquals('Harvest <em class="placeholder">test</em>', $title);
-  }
-
   private function getCommonMockChain() : Chain {
     $options = (new Options())
       ->add('dkan.harvest.service', Harvest::class)
